# github-output: remove chars replacement