--- a/src/set-output.sh
+++ b/src/set-output.sh
@@ -4,4 +4,8 @@ set -e
 
 name="$1"
 value="$2"
-echo "${name}=${value}" >> "$GITHUB_OUTPUT"
+{
+    echo "${name}<<EOF"
+    echo "$value"
+    echo 'EOF'
+} >> "$GITHUB_OUTPUT"