### What

Remove encoding replacement and [use mutitline strings](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#multiline-strings)

### Why

I've just found that the changes in https://github.com/WillAbides/benchdiff-action/pull/4

required some other changes regarding the string manipulation for some special chars (`/r`, `/n`)

See https://github.com/elastic/apm-agent-go/pull/1541#issuecomment-1818398896 that contains what happened


### Test

I used the below workflow to validate the multiline string and no replacements worked fine:

```yaml
---
name: test

on:
  workflow_dispatch: ~

jobs:
  test:
    runs-on: ubuntu-latest
    steps:
      - run: |-
          {
            echo 'pkg: go.elastic.co/apm/v2'
            echo 'goos: linux'
            echo ''
            echo '|      name      | old time/op (ns/op) |  ±  | new time/op (ns/op) |  ±  | delta |        ±        |'
            echo '|----------------|--------------------:|-----|--------------------:|-----|-------|-----------------|'
            echo '| BodyCapturer   |              53.208 | 2%  |              52.746 | 0%  | ~     | (p=0.246 n=5+5) |'
            echo '| BodyCapturer-2 |               52.61 | 1%  |              53.096 | 2%  | ~     | (p=0.421 n=5+5) |'
            echo ' | Transaction    |               984.8 | 12% |              944.36 | 3%  | ~     | (p=0.310 n=5+5) |'
            echo ' | Transaction-2  |               744.8 | 6%  |              806.12 | 24% | ~     | (p=1.000 n=5+5) |'
            echo ''
            echo '|      name      | old alloc/op (B/op) | ±  | new alloc/op (B/op) | ±  | delta |      ±      |'
            echo '|----------------|--------------------:|----|--------------------:|----|-------|-------------|'
            echo '| BodyCapturer   |                   1 | 0% |                   1 | 0% | ~     | (all equal) |'
          } > bench.txt

      - id: benchdiff
        run: |-
          {
            echo "benchstat_output<<EOF"
            cat bench.txt
            echo 'EOF'
          } >> "$GITHUB_OUTPUT"

      - run: |
          {
            echo "## Benchdiff Results"
            echo ""
            echo "${{ steps.benchdiff.outputs.benchstat_output }}"
            echo ""
          } >> $GITHUB_STEP_SUMMARY
```

and produced the below summary:

<img width="1198" alt="image" src="https://github.com/WillAbides/benchdiff-action/assets/2871786/5aceb701-4c7c-4ec6-bb56-7022a4b81718">