@v1v 2023-11-20
https://github.com/elastic/apm-agent-go/actions/runs/6931032834?pr=1543#summary-18851890404 validated my changes