@WillAbides 2023-11-22
After merging this I saw that it breaks the [check-run output](https://github.com/WillAbides/benchdiff-action/runs/18937763554) when the `report_status` option is set. Before creating a release, I'm going to work on finding something that works for both check-run and a job summary. This action predates GITHUB_STEP_SUMMARY, and creating the check-run was a hack around that. So if I can't find something that works for both, I'll just deprecate the check-run and replace it with a job summary.