@WillAbides 2023-11-22
Nevermind what I said above. I forgot that the integration test used the last release instead of HEAD. I fixed that and released [v0.3.5](https://github.com/WillAbides/benchdiff-action/releases/tag/v0.3.5).